597PRFINAL/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pyyaml>=5.4.0
matplotlib>=3.5.0
pytest>=7.0.0
pytest-cov>=3.0.0

